refactor of /content folder and all subfolders related to content. Fix naming, replace type with interface, unify naming of all components
