+++ components/content/ContentDetails.tsx
@@ -1,7 +1,4 @@
 'use client';
-import ParseHtml from '@/components/ParseHtml/ParseHtml';
-import AudioPlayer from '@/components/audioPlayer/AudioPlayer';
-import Comments from '@/components/comments/Comments';
 import { Button } from '@/components/ui/button';
 import { IComment } from '@/lib/validation';
 import { EContentType, IContent } from '@/types/content';
@@ -13,39 +10,35 @@ import {
   Portal,
   Trigger,
 } from '@radix-ui/react-dropdown-menu';
-import { EditIcon } from 'lucide-react';
+import EditIcon from '../icons/Edit';
+import AudioPlayer from './AudioPlayer';
+import Comments from './Comments';
+import HtmlParser from './HtmlParser';
 
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
+import BadgeItem from '../shared/BadgeItem';
 
-type TagProps = {
-  id: string;
-  title: string;
-};
+// ----------------------------------------------------------------
 
-type ContentDetailsProps = {
+interface IContentDetailsProps {
   content: IContent;
-  commentAuthorId: string;
-  contentId: IContent;
-  allComments: IComment[];
-  getAuthorDetails: {
-    isFollowing: boolean;
-    user: IProfileUser;
-  };
-};
+  viewerId: string;
+  comments: IComment[];
+  author: IProfileUser;
+}
 
-const ContentDetails = ({
+const ContentDetails: React.FC<IContentDetailsProps> = ({
   content,
-  contentId,
-  commentAuthorId,
-  getAuthorDetails,
-  allComments,
-}: ContentDetailsProps) => {
+  viewerId,
+  author,
+  comments,
+}) => {
   const router = useRouter();
 
   return (
     <div className="main-content p1-bold w-full mx-auto md:mx-0  space-y-5">
-      {content?.type === EContentType.PODCAST && (
+      {content.type === EContentType.PODCAST && (
         <AudioPlayer
           audioSrc={content.podcastFile!}
           title={content.title}
@@ -53,25 +46,31 @@ const ContentDetails = ({
           coverImage={content.coverImage!}
         />
       )}
-      {content?.type !== EContentType.PODCAST && (
+      {content.type !== EContentType.PODCAST && content.coverImage ? (
         <Image
-          src={content?.coverImage || ` /assets/images/meetup-example.svg`}
+          src={content?.coverImage}
           width={100}
           height={100}
           layout="responsive"
-          alt="post example"
+          alt={content.title}
           className="rounded-2xl w-full min-h-[232px] max-h-[300px] object-cover"
         />
+      ) : (
+        <div className="rounded-2xl w-full min-h-[232px] max-h-[300px] flex-center dark:bg-black-700 bg-primary-100">
+          <Image
+            src="/assets/icons/image-preview.svg"
+            width={100}
+            height={100}
+            alt={content.title}
+          />
+        </div>
       )}
       <div className="flex items-center justify-between">
-        {content?.type === EContentType.MEETUP && (
+        {content.type === EContentType.MEETUP && (
           <>
             <div className="flex gap-2 max-w-3xl">
               <Image
-                src={
-                  getAuthorDetails.user?.avatarImg ||
-                  '/assets/images/post-example.svg'
-                }
+                src={author?.avatarImg || '/assets/images/post-example.svg'}
                 alt="avatar"
                 width={72}
                 height={72}
@@ -86,10 +85,10 @@ const ContentDetails = ({
 
         {content?.type !== EContentType.MEETUP && (
           <h2 className="d2-bold text-wrap break-words overflow-wrap break-word overflow-hidden">
-            {content?.title || 'Title'}
+            {content.title}
           </h2>
         )}
-        {content?.authorId === commentAuthorId && (
+        {content.authorId === viewerId && (
           <DropdownMenu>
             <Trigger asChild>
               <Button className="w-auto">
@@ -108,19 +107,22 @@ const ContentDetails = ({
                 sideOffset={8}
                 align="end"
                 onCloseAutoFocus={(e) => e.preventDefault()}
-                className="bg-light100__dark800 px-2 !w-48 shadow-header-menu data-[side=bottom]:animate-slideUpAndFade data-[side=left]:animate-slideRightAndFade data-[side=right]:animate-slideLeftAndFade data-[side=top]:animate-slideDownAndFade z-20 mb-4 flex  flex-col gap-2.5 rounded-[10px] py-4 ">
+                className="bg-light100__dark800 px-2 !w-48 shadow-header-menu data-[side=bottom]:animate-slideUpAndFade data-[side=left]:animate-slideRightAndFade data-[side=right]:animate-slideLeftAndFade data-[side=top]:animate-slideDownAndFade z-20 mb-4 flex  flex-col gap-2.5 rounded-[10px] py-4 "
+              >
                 <Item
                   onSelect={() => router.push(`/content/${content?.id}/edit`)}
-                  className="flex items-center gap-2.5 focus-visible:outline-none p3-medium cursor-pointer px-3 py-1 rounded-md hover:bg-white-300/30">
+                  className="flex items-center gap-2.5 focus-visible:outline-none p3-medium cursor-pointer px-3 py-1 rounded-md hover:bg-white-300/30"
+                >
                   <EditIcon />
                   <p>Edit Post</p>
                 </Item>
                 <Item
                   onSelect={(e) => e.preventDefault()}
-                  className="flex items-center focus-visible:outline-none  gap-2.5 p3-medium px-3 py-1.5 !text-[#FF584D] hover:bg-white-300/30 rounded-md cursor-pointer">
+                  className="flex items-center focus-visible:outline-none  gap-2.5 p3-medium px-3 py-1.5 !text-error-text hover:bg-white-300/30 rounded-md cursor-pointer"
+                >
                   <Image
                     src="/assets/icons/trash.svg"
-                    width={20}
+                    width={18}
                     height={18}
                     alt="Delete"
                   />
@@ -133,20 +135,16 @@ const ContentDetails = ({
       </div>
       <div>
         <ul className="flex flex-wrap gap-2">
-          {content?.tags.map((tag: TagProps) => (
-            <li
-              key={tag.id}
-              className="dark:bg-black-700 bg-gray-100 rounded-full px-2 py-1 cap-10 uppercase">
-              {tag.title}
-            </li>
+          {content?.tags.map(({ id, title }) => (
+            <BadgeItem key={id} title={title} />
           ))}
         </ul>
       </div>
       <div className="break-words">
-        <ParseHtml data={content?.description} />
+        <HtmlParser data={content?.description} />
       </div>
 
-      {content?.type === EContentType.MEETUP && (
+      {content.type === EContentType.MEETUP && (
         <>
           <div className="flex gap-2">
             <Image
@@ -178,9 +176,9 @@ const ContentDetails = ({
         </>
       )}
       <Comments
-        contentId={contentId.id}
-        commentAuthorId={commentAuthorId}
-        allComments={allComments}
+        contentId={content.id}
+        commentAuthorId={viewerId}
+        comments={comments}
       />
     </div>
   );
