+++ components/layout/Header.tsx
@@ -6,6 +6,7 @@ import NavMenu from '../navigation/NavMenu';
 import Image from 'next/image';
 
 import { useTheme } from '@/context/ThemeProvider';
+import Link from 'next/link';
 
 // ----------------------------------------------------------------
 
@@ -21,7 +22,9 @@ const Header: React.FC<IHeaderProps> = (props) => {
 
   return (
     <header className="bg-light100__dark800 flex-between fixed left-0 top-0 z-50 w-full gap-5 px-3.5 py-4 lg:px-8 lg:py-5">
-      <Image src={logoUrl} width={147} height={30} alt="Logo" />
+      <Link href="/" className="w-[147px] h-[30px] relative">
+        <Image src={logoUrl} fill alt="Logo" />
+      </Link>
       <div className="max-sm:hidden">
         <NavLinks />
       </div>
